move TokenValidator into internal/api
It is only needed here, so it should not be constructed if we don't need
it (esp. during test-get-quota etc.).
