--- internal/api/token.go
@@ -1,36 +1,0 @@
-/*******************************************************************************
-*
-* Copyright 2017 SAP SE
-*
-* Licensed under the Apache License, Version 2.0 (the "License");
-* you may not use this file except in compliance with the License.
-* You should have received a copy of the License along with this
-* program. If not, you may obtain a copy of the License at
-*
-*     http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*
-*******************************************************************************/
-
-package api
-
-import (
-	"net/http"
-
-	"github.com/gorilla/mux"
-	"github.com/sapcc/go-bits/gopherpolicy"
-)
-
-// CheckToken checks the validity of the request's X-Auth-Token in Keystone, and
-// returns a Token instance for checking authorization. Any errors that occur
-// during this function are deferred until Require() is called.
-func (p *v1Provider) CheckToken(r *http.Request) *gopherpolicy.Token {
-	t := p.Cluster.Auth.TokenValidator.CheckToken(r)
-	t.Context.Request = mux.Vars(r)
-	return t
-}
